feat: email ui 중 footer 컴포넌트화
